requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
requests>=2.28.0
python-dotenv>=1.0.0
instaloader>=4.15
streamlit>=1.28.0
selenium>=4.0.0
